// convex/_generated/server.d.ts
/* prettier-ignore-start */

/* eslint-disable */
/**
 * Generated server utility.
 *
 * THIS CODE IS AUTOGENERATED, DO NOT EDIT.
 *
 * To update this file, run `npx convex dev`.
 *
 * @module
 */

import type {
  GenericQueryCtx,
  GenericMutationCtx,
  GenericActionCtx,
  GenericDatabaseReader,
  GenericDatabaseWriter,
} from "convex/server";
import type { DataModel } from "./dataModel";

/**
 * Defines a helper to inject a database into a function.
 */
export declare const query: <Args, Output>(
  queryFn: (ctx: GenericQueryCtx<DataModel>, args: Args) => Output
) => (ctx: GenericQueryCtx<DataModel>, args: Args) => Output;
export declare const mutation: <Args, Output>(
  mutationFn: (ctx: GenericMutationCtx<DataModel>, args: Args) => Output
) => (ctx: GenericMutationCtx<DataModel>, args: Args) => Output;
export declare const action: <Args, Output>(
  actionFn: (ctx: GenericActionCtx<DataModel>, args: Args) => Output
) => (ctx: GenericActionCtx<DataModel>, args: Args) => Output;

/* prettier-ignore-end */